backend/pyproject.toml
[tool.poetry]
name = "Accessible transport"
version = "0.1.0"
description = ""
authors = ["Камбаллисты teams"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
flake8 = "^7.0.0"
drf-yasg = "^1.21.7"
networkx = "^3.3"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
